requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pydantic==2.3.0
pypdf2==3.0.1
langchain==0.0.312
langchain-openai==0.0.2
openai
tiktoken
faiss-cpu==1.7.4
jinja2==3.1.2
python-dotenv==1.0.0
requests==2.31.0 